refactored analysisService related classes into analysis package
