+++ song-server/src/main/java/bio/overture/song/server/service/DonorService.java
@@ -31,6 +31,7 @@ import static com.google.common.collect.ImmutableList.toImmutableList;
 import bio.overture.song.server.model.entity.Donor;
 import bio.overture.song.server.model.entity.composites.DonorWithSpecimens;
 import bio.overture.song.server.repository.DonorRepository;
+import bio.overture.song.server.service.InfoService.DonorInfoService;
 import bio.overture.song.server.service.id.IdService;
 import java.util.ArrayList;
 import java.util.List;
