+++ song-server/src/main/java/bio/overture/song/server/service/ExportService.java
@@ -34,6 +34,7 @@ import bio.overture.song.server.model.entity.Donor;
 import bio.overture.song.server.model.entity.FileEntity;
 import bio.overture.song.server.model.entity.Sample;
 import bio.overture.song.server.model.entity.Specimen;
+import bio.overture.song.server.service.analysis.AnalysisService;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableList;
