+++ song-server/src/main/java/bio/overture/song/server/service/SampleService.java
@@ -32,6 +32,7 @@ import bio.overture.song.server.model.entity.BusinessKeyView;
 import bio.overture.song.server.model.entity.Sample;
 import bio.overture.song.server.repository.BusinessKeyRepository;
 import bio.overture.song.server.repository.SampleRepository;
+import bio.overture.song.server.service.InfoService.SampleInfoService;
 import bio.overture.song.server.service.id.IdService;
 import java.util.List;
 import javax.transaction.Transactional;
