+++ song-server/src/test/java/bio/overture/song/server/controller/AnalysisControllerTest.java
@@ -42,9 +42,9 @@ import bio.overture.song.core.utils.ResourceFetcher;
 import bio.overture.song.server.model.analysis.Analysis;
 import bio.overture.song.server.model.dto.Payload;
 import bio.overture.song.server.model.dto.schema.RegisterAnalysisTypeRequest;
-import bio.overture.song.server.service.AnalysisService;
-import bio.overture.song.server.service.AnalysisServiceSender;
 import bio.overture.song.server.service.StudyService;
+import bio.overture.song.server.service.analysis.AnalysisService;
+import bio.overture.song.server.service.analysis.AnalysisServiceSender;
 import bio.overture.song.server.utils.EndpointTester;
 import bio.overture.song.server.utils.generator.AnalysisGenerator;
 import bio.overture.song.server.utils.generator.StudyGenerator;
