+++ song-server/src/test/java/bio/overture/song/server/controller/EnforcedSubmitControllerTest.java
@@ -30,8 +30,8 @@ import static junit.framework.TestCase.assertTrue;
 import bio.overture.song.core.model.AnalysisTypeId;
 import bio.overture.song.core.utils.ResourceFetcher;
 import bio.overture.song.server.model.dto.UpdateAnalysisRequest;
-import bio.overture.song.server.service.AnalysisService;
 import bio.overture.song.server.service.StudyService;
+import bio.overture.song.server.service.analysis.AnalysisService;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.nio.file.Paths;
 import lombok.NonNull;
