+++ song-server/src/test/java/bio/overture/song/server/service/FileModificationServiceTest.java
@@ -53,6 +53,7 @@ import bio.overture.song.core.utils.Streams;
 import bio.overture.song.server.converter.FileConverter;
 import bio.overture.song.server.model.entity.FileEntity;
 import bio.overture.song.server.repository.FileRepository;
+import bio.overture.song.server.service.analysis.AnalysisServiceImpl;
 import java.util.List;
 import java.util.Set;
 import java.util.function.BiConsumer;
