+++ song-server/src/test/java/bio/overture/song/server/service/InfoServiceTest.java
@@ -28,6 +28,7 @@ import static org.junit.Assert.assertTrue;
 import bio.overture.song.core.testing.SongErrorAssertions;
 import bio.overture.song.core.utils.JsonUtils;
 import bio.overture.song.core.utils.RandomGenerator;
+import bio.overture.song.server.service.InfoService.DonorInfoService;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
