+++ song-server/src/test/java/bio/overture/song/server/service/StudyWithDonorsServiceTest.java
@@ -31,6 +31,7 @@ import bio.overture.song.server.model.entity.Specimen;
 import bio.overture.song.server.model.entity.composites.CompositeEntity;
 import bio.overture.song.server.model.entity.composites.DonorWithSpecimens;
 import bio.overture.song.server.model.entity.composites.SpecimenWithSamples;
+import bio.overture.song.server.service.analysis.AnalysisService;
 import bio.overture.song.server.utils.generator.StudyGenerator;
 import com.google.common.collect.Maps;
 import java.util.Collection;
