+++ song-server/src/test/java/bio/overture/song/server/service/SubmitServiceTest.java
@@ -41,6 +41,7 @@ import bio.overture.song.core.utils.RandomGenerator;
 import bio.overture.song.core.utils.Responses;
 import bio.overture.song.server.model.dto.Payload;
 import bio.overture.song.server.repository.UploadRepository;
+import bio.overture.song.server.service.analysis.AnalysisService;
 import bio.overture.song.server.service.id.IdService;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
